renamed internal/testutil to internal/dump, removed requests parsing
