+++ internal/dump/dump.go
@@ -1,23 +1,23 @@
-package testutil
+package dump
 
 import (
 	"github.com/indigo-web/indigo/http"
 	"strconv"
 )
 
-func SerializeRequest(request *http.Request) (string, error) {
+func Request(request *http.Request) (string, error) {
 	var buff []byte
 
 	buff = append(buff, request.Method.String()...)
-	buff = space(buff)
+	buff = append(buff, ' ')
 	buff = append(buff, request.Path...)
 
 	if !request.Params.Empty() {
-		buff = question(buff)
+		buff = append(buff, '?')
 
 		for key, value := range request.Params.Iter() {
 			buff = queryparam(buff, key, value)
-			buff = ampersand(buff)
+			buff = append(buff, '&')
 		}
 
 		if len(buff) > 0 && buff[len(buff)-1] == '&' {
@@ -25,48 +25,42 @@ func SerializeRequest(request *http.Request) (string, error) {
 		}
 	}
 
-	buff = space(buff)
+	buff = append(buff, ' ')
 	buff = append(buff, request.Protocol.String()...)
-	buff = crlf(buff)
+	buff = append(buff, '\r', '\n')
 
 	for _, h := range request.Headers.Expose() {
 		buff = header(buff, h)
 	}
 
-	buff = header(buff, http.Header{
-		Key:   "Content-Length",
-		Value: strconv.Itoa(request.ContentLength),
-	})
+	if request.Encoding.Chunked {
+		buff = append(buff, "Transfer-Encoding: "...)
+		for _, enc := range request.Encoding.Transfer {
+			buff = append(buff, enc...)
+			buff = append(buff, ',', ' ')
+		}
 
-	buff = crlf(buff)
+		buff = append(buff, "chunked\r\n"...)
+	} else {
+		buff = header(buff, http.Header{
+			Key:   "Content-Length",
+			Value: strconv.Itoa(request.ContentLength),
+		})
+	}
+
+	buff = append(buff, '\r', '\n')
 	body, err := request.Body.Bytes()
 	buff = append(buff, body...)
 
 	return string(buff), err
 }
 
-func space(b []byte) []byte {
-	return append(b, ' ')
-}
-
-func question(b []byte) []byte {
-	return append(b, '?')
-}
-
-func ampersand(b []byte) []byte {
-	return append(b, '&')
-}
-
-func crlf(b []byte) []byte {
-	return append(b, '\r', '\n')
-}
-
 func header(b []byte, h http.Header) []byte {
 	b = append(b, h.Key...)
-	b = colonsp(b)
+	b = append(b, ':', ' ')
 	b = append(b, h.Value...)
 
-	return crlf(b)
+	return append(b, '\r', '\n')
 }
 
 func queryparam(buff []byte, key, value string) []byte {
@@ -74,8 +68,4 @@ func queryparam(buff []byte, key, value string) []byte {
 	buff = append(buff, '=')
 	buff = append(buff, value...)
 	return buff
-}
-
-func colonsp(b []byte) []byte {
-	return append(b, ':', ' ')
 }
